fixed study-room UI
