+++ Server/.env
@@ -1,6 +1,5 @@
 JWT_SECRET = manishkp842925
-# MONGODB_URI = mongodb+srv://manishkp842:jMOEQZdQuhw6pgds@project.ung8d.mongodb.net/GroupProject?retryWrites=true&w=majority&appName=project
+MONGODB_URI = mongodb+srv://manishkp842:jMOEQZdQuhw6pgds@project.ung8d.mongodb.net/GroupProject?retryWrites=true&w=majority&appName=project
 
-# MONGODB_URI =mongodb://rishukumar9233:K0UhYj92G9KPEq8s@undefined/?replicaSet=atlas-x53f46-shard-0&ssl=true&authSource=admin
-MONGODB_URI = mongodb+srv://rishukumar9233:K0UhYj92G9KPEq8s@cluster0.gqce3.mongodb.net/?retryWrites=true&w=majority&appName=Cluster0
+# MONGODB_URI = mongodb+srv://rishukumar9233:K0UhYj92G9KPEq8s@cluster0.gqce3.mongodb.net/?retryWrites=true&w=majority&appName=Cluster0
